--- src/devices/SM28BYJ48/SM28BYJ48.csproj
@@ -1,19 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-    <!--Disabling default items so samples source won't get build by the main library-->
-    <EnableDefaultItems>false</EnableDefaultItems>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="System.Device.Gpio" Version="0.1.0-prerelease*" />
-    <Compile Include="SM28BYJ48.cs" />
-  </ItemGroup>
-
-  <ItemGroup> 
-    <None Remove="README.md" />
-    <None Remove="SM28BYJ48.png" />
-  </ItemGroup>
-
-</Project>
