+++ src/devices/StepperMotor/StepperMotor.csproj
@@ -1,22 +1,20 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp2.1</TargetFramework>
+    <!--Disabling default items so samples source won't get build by the main library-->
+    <EnableDefaultItems>false</EnableDefaultItems>
   </PropertyGroup>
 
- <ItemGroup>
+  <ItemGroup>
     <PackageReference Include="System.Device.Gpio" Version="0.1.0-prerelease*" />
+    <Compile Include="StepperMotor.cs" />
+    <Compile Include="StepperMode.cs" />
   </ItemGroup>
 
- <ItemGroup>
-    <ProjectReference Include="$(MSBuildThisFileDirectory)/../SM28BYJ48.csproj" />
-  </ItemGroup>
-
-    <ItemGroup> 
+  <ItemGroup> 
     <None Remove="README.md" />
     <None Remove="SM28BYJ48.png" />
-    <None Remove="SM28BYJ48.fzz" />
   </ItemGroup>
 
 </Project>
