+++ src/devices/StepperMotor/samples/StepperMotor.Samples.csproj
@@ -1,0 +1,22 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
+  </PropertyGroup>
+
+ <ItemGroup>
+    <PackageReference Include="System.Device.Gpio" Version="0.1.0-prerelease*" />
+  </ItemGroup>
+
+ <ItemGroup>
+    <ProjectReference Include="$(MSBuildThisFileDirectory)/../StepperMotor.csproj" />
+  </ItemGroup>
+
+  <ItemGroup> 
+      <None Remove="README.md" />
+      <None Remove="SM28BYJ48.png" />
+      <None Remove="SM28BYJ48.fzz" />
+  </ItemGroup>
+
+</Project>
